requirements.txt
Jinja2
openpyxl
pandas
python-binance
schedule
streamlit
streamlit-extras
Pillow
plotly
validators
yfinance
Markdown
markdown2
pdfkit